[1.14] Fix setting etcd-server when no mTLS is set between etcd and kube-apiserver
An error is introduced in commit f0066e817575ed09d6870943f1d52e8e041800d7 when I cherry pick https://github.com/kubernetes/kubernetes/pull/74690, when mTLS is not enabled, etcd-server was not set, this patch is to fix the issue. This PR is temporarily rolled back in OSS for performance issue investigation, so I cannot fix it in OSS now. I am fixing this in GKE for now, I will add a buganizer to track and make sure when this is fixed in open source, this fix will not carry on.

Change-Id: I0a13e91761576993f5f8f92bdaec2fa375149e7c