Fix string assertion in vsphere E2E test
This test asserts a hard-coded error string, which doesn't actually
match the error produced by production code. It is unfortunate to have
such a brittle test, but we don't want to spend time investing in this
particular test suite.

As it happens, this string was actually fixed as a minor change in a
completely unrelated PR: https://github.com/kubernetes/kubernetes/pull/79931

In order to avoid pulling in lots of unrelated changes, I am making this
fix directly.

Tested: on my own anthos cluster
Bug: 145919661
Change-Id: Idbc26994c23e0bd8c9ab3bcf8bc1223c4ec315b0